+++ src/components/auth/logo/Logo.jsx
